Using paths instead of circles to support pie charts in one of the charts
Updated the bubble chart and the scatterplot to work as arguments to
.call, rather than passing in the container to which they add
themselves. This helps a bit with maintaining data constancy because I
can create a bunch of g elements with the data returned from the server
and then just .call the different charts on that selection.

Moved all three charts to paths instead of circles so that the small
multiples can display pie charts instead of just circles so you can see
the breakdown of costs and benefits for each demographic. Eventually I
should be able to add a custom tween function to animate the
transitions between charts.