Updated how data is handled by server
Make the incentives map/reduce functions index on the sector and return
the entire document. Rather than joining public/private data in the
server with a list, just grab the private data and simplify the data
structure on the client-side. Still need to update the pie charts and
the scatterplot to handle this change.